Enable more reliable source compilation on Windows
Windows does not have a standard way to define the location of headers
and libraries like Linux. Multiple solutions exist (vcpkg, conan, ...)
that attempt to fix this, but they rely on different mechanisms. This PR
tries to add a bit more flexibility to this process.

This PR makes it so that a user not using Anaconda (for whatever reason)
is still able to successfully compile the library and the Python module.
Right now, it's not always the case that `find_package` succedes: e.g.
when using `vcpkg` to handle C/C++ dependencies, the linker arguments
supplied will make the build fail. I do not think this is a `gdstk` bug,
that's why I did not create an issue.

It uses `find_package` by default, so that Linux builds are not affected
but on failure to locate the library, two mecahnisms are tried:

1. The user can set `ZLIB_INCLUDE_DIRS` and `ZLIB_LIBRARIES` as
   as environment variables, and they will be used **if `find_package`
   fails**. If `ZLIB_INCLUDE_DIRS` is not set, the build will fail.
2. If `ZLIB_LIBRARIES` is not set, then `find_library` will be used,
   which should scan tha `PATH` variable on Windows for possible
   libraries.
3. The build fails if all these three mechanisms fail, in this order:
   `find_package`, unset `ZLIB_LIBRARY`, `find_library`.