Fix ref counting in `Reference::apply_repetition`
The reference count of the referenced (raw) cells needs to be increased
for each created Reference. Otherwise, DECREF in reference_object_dealloc is not balanced.

This issue becomes visible when the return value of `apply_repetition`
is not assigned, e.g. by calling `ref.apply_repetition()` instead of
`out = ref.apply_repetition()`, as this results in immediate deallocation of
the returned list and its elements.